+++ src/r0/contact.rs
@@ -1,5 +1,4 @@
 //! Endpoints for account contact information.
 
-pub mod create_contact;
 pub mod get_contacts;
 pub mod request_contact_verification_token;