+++ src/r0/session.rs
@@ -4,3 +4,4 @@ pub mod get_login_types;
 pub mod login;
 pub mod logout;
 pub mod logout_all;
+pub mod sso_login;