+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Improvements:
 * Add an `Error` type that represents the well-known errors in the client-server API
 * Add OpenID token request endpoint.
 * Add `r0::client_exchange::send_event_to_device` (introduced in r0.3.0)
+* Add endpoints to retrieve account_data (introduced in r0.5.0)
 
 Breaking changes: