+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Breaking changes:
 * Update ruma-api to 0.14.0
 * Fix `r0::session::get_login_types`
 * Add `allow_remote` parameter to `r0::media::get_content`
+* Add missing parameters for `r0::room::create_room`
 
 # 0.6.0