+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # [unreleased]
 
+# 0.17.0
+
 Breaking changes:
 
 * `collections::only` no longer exports a `raw` submodule. It was never meant ot be exported in the first place.